Refactored IWatcherService into a single file
